Refactored 'inputs' package to 'library'
