+++ src/org/usfirst/frc/team4624/robot/subsystems/Powertrain.java
@@ -3,7 +3,7 @@ package org.usfirst.frc.team4624.robot.subsystems;
 
 import org.usfirst.frc.team4624.robot.RobotMap;	// ENUMS for the ports
 import org.usfirst.frc.team4624.robot.commands.DriveCommand;
-import org.usfirst.frc.team4624.robot.input.XboxController;
+import org.usfirst.frc.team4624.robot.library.XboxController;
 import edu.wpi.first.wpilibj.Jaguar;
 import edu.wpi.first.wpilibj.RobotDrive;
 import edu.wpi.first.wpilibj.command.Subsystem;
